% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-answer-block.R
\name{answer_style}
\alias{answer_style}
\title{Whether to display answers.}
\usage{
answer_style(x, remember = TRUE)
}
\description{
answer_style() gives the currently active answer style
and allows it to become the default if remember=TRUE
pop_answer_style() reverts to the previous answer style
}
