% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{check_submission_names}
\alias{check_submission_names}
\title{Are the submissions names a subset of the class roster. If not,
that suggests that either the submissions or the class roster are wrong.}
\usage{
check_submission_names(
  home = ".",
  since = "2000-1-1 00:00:01 UTC",
  silent = FALSE
)
}
\description{
Are the submissions names a subset of the class roster. If not,
that suggests that either the submissions or the class roster are wrong.
}
