% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grading_directory.R
\name{create_grading_directory}
\alias{create_grading_directory}
\title{Create the stub of a devoirs grading directory}
\usage{
create_grading_directory(course_name)
}
\arguments{
\item{course_name}{A string specifying the name of the course.}
}
\description{
It's recommended that you keep `course_name` short and avoid any white space or
characters that are disallowed in file or directory names. Example: "Stats_101_Fall_2025".
}
