% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC_score_table.R
\name{create_MC_score_table}
\alias{create_MC_score_table}
\title{Create a scoring table for MC questions in a document}
\usage{
create_MC_score_table(Submissions, min_count = 0, min_frac = 0)
}
\description{
Run `summarize_document(docid, since, until)` to get list containing the MC
raw scores, the ungraded essays, and the ungraded R elements (checked for parsing)
}
