% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devoirs_memory.R
\docType{data}
\name{devoirs_memory}
\alias{devoirs_memory}
\title{Store question labels to avoid repeats}
\format{
An object of class \code{environment} of length 1.
}
\usage{
devoirs_memory
}
\description{
Store question labels to avoid repeats
}
\keyword{datasets}
