% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devoirs_set.R
\name{devoirs_set}
\alias{devoirs_set}
\title{Make a multiple choice for a small set}
\usage{
devoirs_set(
  choices = 1:5,
  correct = 3,
  n = 5,
  id = NULL,
  sort = is.numeric(correct),
  inline = TRUE
)
}
\arguments{
\item{choices}{Vector containing the member of the set to use as distractors}

\item{correct}{Number or character string with the <correct> answer.}

\item{n}{integer, number of distractors}

\item{id}{character string. File-unique id for this problem.}

\item{sort}{logical value. Force sorting on character strings. (Numerical choices are always sorted.)}

\item{inline}{logical value. Place the items in an inline format.}
}
\description{
Make a multiple choice for a small set
}
