% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devoirs_start.R
\name{devoirs_start}
\alias{devoirs_start}
\title{Placement of document objects needed to collect answers}
\usage{
devoirs_start(documentID = knitr::current_input(), collectURL = NULL)
}
\arguments{
\item{documentID}{String with a unique ID.}

\item{collectURL}{String containing URL where students can submit
JSON holding their work.}
}
\description{
Placement of document objects needed to collect answers
}
