% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devoirs-text-entry.R
\name{devoirs_text}
\alias{devoirs_text}
\title{Create a text-entry widget for collection by devoirs package}
\usage{
devoirs_text(
  qID,
  rows = 4,
  cols = 50,
  contents = "",
  code = NULL,
  str = "Enter your response to {code} here"
)
}
\arguments{
\item{qID}{unique text string}

\item{rows}{number of lines in empty text area (4)}

\item{cols}{number of characters available in each line (50)}

\item{contents}{Character string which will become the default contents of the widget.}

\item{str}{Sets the prompt in the box that disappears}
}
\description{
Create a text-entry widget for collection by devoirs package
}
