% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReportsV2.R
\name{submissions_by_student}
\alias{submissions_by_student}
\alias{MC_items}
\title{Reports for Version 2 of the grading system}
\usage{
submissions_by_student(home = ".")

MC_items(home = ".", Items = NULL)
}
\description{
There must be an All_items.RSD in the grading directory.
If not, run update_items() to create one.
}
