% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true-or-false.R
\name{devoirsTF}
\alias{devoirsTF}
\title{True or false question}
\usage{
devoirsTF(
  correct = FALSE,
  id = NULL,
  wrong.hint = "",
  right.hint = "",
  leed = "",
  format = "True or False"
)
}
\arguments{
\item{correct}{logical TRUE or FALSE indicating which is correct}

\item{id}{an alphanumerical id. Will be assigned randomly if not specified, but this is
not recommended.}

\item{wrong.hint}{The hint (if any) to be given for a wrong response. Keep in mind that the
wrong response is determined by the `correct` argument}

\item{right.hint}{The hint for a correct response.}

\item{leed}{String to preceed the radio buttons}

\item{format}{String like "yes or no?" Default: "True or False"}
}
\description{
A compact option for authors writing lots of true-or-false questions.
}
\details{
The leftmost token (without punctuation) of the leed becomes the label for the left
radio button. Similarly, the rightmost token will label the right radio button.
}
