% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{format_R}
\alias{format_R}
\title{Score multiple choice problems for one student for one assignment
THIS HAS BEEN REPLACED by version in score_MC.R
Format the R history for display
Print this with knitr::kable()
Maybe reformat <time> to be "days ago"}
\usage{
format_R(Revents, time_unit = NULL)
}
\description{
Score multiple choice problems for one student for one assignment
THIS HAS BEEN REPLACED by version in score_MC.R
Format the R history for display
Print this with knitr::kable()
Maybe reformat <time> to be "days ago"
}
