% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{get_class_roster}
\alias{get_class_roster}
\title{Who is registered for the class}
\usage{
get_class_roster(home = ".", section = NULL, silent = FALSE)
}
\arguments{
\item{section}{FOR FUTURE USE}
}
\description{
Who is registered for the class
}
