% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_course_params.R
\name{get_course_params}
\alias{get_course_params}
\title{Get parameters for a devoirs-graded course directory}
\usage{
get_course_params(home = ".", silent = FALSE)
}
\arguments{
\item{home}{Character string specifying the grading directory for the course}

\item{silent}{Logical flag: give a message.}
}
\description{
Parses the YAML file with the course information and
re-arranges things like the student list to be primary and alias names
}
