% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{get_course_params}
\alias{get_course_params}
\alias{get_new_submissions}
\alias{get_historic_data}
\alias{update_submissions}
\alias{document_names}
\title{Functions for grading}
\usage{
get_course_params(home = ".", silent = FALSE)

get_new_submissions(home = ".", silent = FALSE)

get_historic_data(home = ".", since = "2000-1-1 00:00:01 UTC", silent = FALSE)

update_submissions(home = ".", silent = FALSE)

document_names(home = ".", since = "2000-1-1 00:00:01 UTC")
}
\arguments{
\item{home}{The directory to use. Default: the current working directory.}

\item{since}{date for earliest submission to include. Format: "2024-11-18 00:00:01 UTC"}

\item{new_only}{If TRUE, just return the data newly read from the repo.}
}
\description{
Assumptions: All work is done in a custom-made directory, just for the course
That directory has a file course-parameters.yml that defines the course name and the
file (or Google spreadsheet) that holds the submissions.
}
\details{
Make sure we are in a valid directory and get the params
}
