% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_V2.R
\name{update_items}
\alias{update_items}
\alias{JSON_to_ITEMS}
\title{Command line utilities for a devoirs database}
\usage{
update_items(home = ".")

JSON_to_ITEMS(JSONentries)
}
\arguments{
\item{home}{the name of the home directory, usually "."}

\item{JSONentries}{the raw json as provided by get_rawnew_submissions()
or as recorded in the}
}
\description{
JSON is the material from the upload site
ITEMS is the parsed item-by-item data frame
}
