% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich-webr-chunk.R
\name{insert_rich_webr_chunk}
\alias{insert_rich_webr_chunk}
\title{Insert a webr chunk into a {devoirs} document.}
\usage{
insert_rich_webr_chunk(ename, text_submit = TRUE, caption = "Webr chunk")
}
\arguments{
\item{ename}{Name of the file or other reference name}
}
\description{
Insert a webr chunk into a {devoirs} document.
}
