% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{is_valid_directory}
\alias{is_valid_directory}
\title{Functions for grading}
\usage{
is_valid_directory(home = ".")
}
\description{
Assumptions: All work is done in a custom-made directory, just for the course
That directory has a file course-parameters.yml that defines the course name and the
URL (e.g. via  Google spreadsheet) that holds the submissions.
}
