% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_roster.R
\name{make_roster}
\alias{make_roster}
\alias{roster2catalog}
\title{Create a catalog of the exercise files in a directory.}
\usage{
make_roster(chapter_dir)

roster2catalog(chapter_dir)
}
\arguments{
\item{chapter_dir}{Character string with the name of the already-existing directory
containing the exercise Quarto files.}
}
\description{
For reference purposes, it can be helpful to have a catalog of all
of the exercise files in a directory. These two functions help to do
this. There are two phases:
\enumerate{
\item Create a YAML roster of all the files. This is done with \code{make_roster()}
The roster lives in the same directory
as the exercise files.
\item Translate the roster file into a Quarto file that will compile to the
catalog. \code{roster2catalog()} handles this phase
}
}
\details{
The roster file contains YAML, one entry for each exercise file. Edit this
by hand to divide the exercises into groups, which will appear as separate tabsets
in the catalog. The \verb{group:} field specifies the group for each exercise. If you add more
exercise files to the directory after building the roster, just build it again. The hand-edited
information will be retained.

Create or refresh a chapter roster while retaining editorial fields.
}
