% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_doc_MC.R
\name{MC_doc_summary}
\alias{MC_doc_summary}
\title{Get a summary of MC scores in a given document}
\usage{
MC_doc_summary(
  home = ".",
  docid,
  Items = valid_All_items(home),
  sections = "All"
)
}
\description{
Get a summary of MC scores in a given document
}
