% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcq_engine.R
\name{mcq_engine}
\alias{mcq_engine}
\title{Facilities to format multiple-choice questions.}
\usage{
mcq_engine(options)
}
\arguments{
\item{options}{A placeholder argument used by knitr.}
}
\description{
Multiple-choice questions are formatted using a knitr chunk of type `{mcq}`
(rather than the ordinary `{r}`). The advantage of using a chunk is that unquoted
text can be given for the question choices. There is a special format for the contents
of an `{mcq}` chunk, described in the vignettes. The exported function `mcq_engine()`
is put into place with `knitr::knit_engines$set("mcq" = mcq_engine)` which is
called (via `.onload()`) when the `{devoirs}` package is attached.
}
