% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_MC.R
\name{merge_scores}
\alias{merge_scores}
\title{Combine the stored scores (old) if any with
scores derived from the raw submissions (new)}
\usage{
merge_scores(old, new, default_score = 2)
}
\description{
Combine the stored scores (old) if any with
scores derived from the raw submissions (new)
}
