% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise-listing.R
\name{new_exercise_listing}
\alias{new_exercise_listing}
\title{Create an exercise listing and template file}
\usage{
new_exercise_listing(ename = new_exercise_name(), dir = "_exercises/")
}
\description{
Create an exercise listing and template file
}
