% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new-exercise-name.R
\name{new_exercise_name}
\alias{new_exercise_name}
\title{Create a random name for an exercise}
\usage{
new_exercise_name()
}
\description{
Functions for writing exercises
}
\details{
Create a random name using the \if{html}{\out{<animal>}}-\if{html}{\out{<verb>}}-\if{html}{\out{<noun>}} pattern.
}
