% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise-listing.R
\name{new_item_label}
\alias{new_item_label}
\title{Make a label for a new devoirs item}
\usage{
new_item_label(ename)
}
\arguments{
\item{ename}{the root name of a file (e.g.
'cat-seeks-mouse') or a full path name.}
}
\description{
Make a label for a new devoirs item
}
