% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade_V2.R
\name{parse_json}
\alias{parse_json}
\title{Translate from a vector of JSON submissions to a list of R lists
One list for each submission, broken down into "essay" and "MC"
(multiple choice) types.}
\usage{
parse_json(contents)
}
\description{
Translate from a vector of JSON submissions to a list of R lists
One list for each submission, broken down into "essay" and "MC"
(multiple choice) types.
}
