% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_doc_MC.R
\name{report_doc_MC}
\alias{report_doc_MC}
\title{Multiple-choice report for one document}
\usage{
report_doc_MC(
  home = ".",
  docid,
  Items = valid_All_items(home),
  sections = "All"
)
}
\arguments{
\item{Items}{data frame of all items in this course,
if not specified, will read from the grading directory.}
}
\description{
Multiple-choice report for one document
}
