% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_one_student.R
\name{report_one_student}
\alias{report_one_student}
\title{Summarize all items for one student}
\usage{
report_one_student(home = ".", student)
}
\description{
A document-by-document summary for that student
}
