% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_document.R
\name{score_document}
\alias{score_document}
\title{Give a comprehensive score on one document for all students}
\usage{
score_document(
  home = ".",
  docid = "Drill-02-quantities",
  essay_weight = 2.5,
  Submissions = get_historic_data(home),
  sections = "All",
  write_csv = FALSE,
  since = "2000-1-1 00:00:01 UTC",
  until = Sys.time() + (24 * 60 * 60 - 1)
)
}
\description{
Give a comprehensive score on one document for all students
}
