% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_MC.R
\name{score_MC}
\alias{score_MC}
\title{Construct MC grading matrices for a document}
\usage{
score_MC(submissions, min_count = 0, min_frac = 0, document = "unknown")
}
\arguments{
\item{submissions}{The MC component of the summarize_document() output}

\item{count}{Minimum count (as integer) for item to be scored}

\item{frac}{Minimum fraction (as decimal) for item to be scored}
}
\description{
Construct MC grading matrices for a document
}
