% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scoring.R
\name{Scoring}
\alias{Scoring}
\alias{ScoreEssays}
\alias{ViewDocument}
\title{Run apps for viewing and scoring answers}
\usage{
ScoreEssays(dir_name = NULL)

ViewDocument()
}
\description{
These apps provide a GUI to view student submissions and score essays.
}
\details{
\code{ViewDocument()} provides an interface to one document at a time. It
consolidates in one place both essay and multiple-choice submissions.

\code{ScoreEssays()} is for assigning a score to an essay submission. It can
move between documents and items within a document. At any instant, the app
displays one student's answer to an essay item so that it can be scored. You
can easily move from student-to-student.

To use these apps you must have a "grading directory" set up for
the course.
}
