% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_doc_MC.R
\name{students_in_section}
\alias{students_in_section}
\title{Get the names of students in the given sections}
\usage{
students_in_section(home = ".", sections = "All")
}
\description{
Get the names of students in the given sections
}
