% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{summarize_document}
\alias{summarize_document}
\title{Summarize submissions from multiple students from a single document}
\usage{
summarize_document(
  home = ".",
  students = get_class_roster(home)$class_list,
  docid = "03-exercises.rmarkdown",
  Submissions = get_historic_data(home),
  since = "2000-1-1 00:00:01 UTC",
  until = Sys.time() + (24 * 60 * 60 - 1)
)
}
\description{
Summarize submissions from multiple students from a single document
}
