% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{summarize_student_doc}
\alias{summarize_student_doc}
\title{Summarize all of a student's submissions from a single file.}
\usage{
summarize_student_doc(
  Submissions = get_historic_data(),
  docid = "03-exercises.rmarkdown",
  student = "dtkaplan@gmail.com",
  since = "2000-1-1 00:00:01 UTC",
  until = Sys.Date() + (24 * 60 * 60 - 1)
)
}
\description{
Summarize all of a student's submissions from a single file.
}
