% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert-answer-block.R
\name{push_answer_style}
\alias{push_answer_style}
\alias{pop_answer_style}
\title{Controls answer/hint display in `{devoirs}` items.}
\usage{
push_answer_style(type)

pop_answer_style(clear = FALSE)
}
\arguments{
\item{type}{The type of display.
Currently only `"none"` (no display) and `"block"` (do display) are allowed.}

\item{remember}{DEPRECATED. Use this `type` for future questions.
Equivalent to `push_answer_style(type)`.}
}
\description{
This is still in draft state and uses a simple CSS display style mechanism.
Optionally, authors using `{devoirs}` can choose to display answers
or hints for multiple-choice questions. The  `answer_style()` functions
described here allow the display to be controlled on a section or document
level.
}
\details{
item-by-item basis with the `#| show_hints: true` chunk option.
The section or document level.



- `push_answer_style()` stores the display type for successive questions.
- `pop_answer_style()` reverts to the previous answer style pushed.
}
