% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_grading_directory.R
\name{create_grading_directory}
\alias{create_grading_directory}
\title{Create the stub of a devoirs grading directory}
\usage{
create_grading_directory(course_name)
}
\arguments{
\item{course_name}{A string specifying the name of the course. This should be a
valid directory name.}
}
\description{
Corresponding to the "collection site" for \code{{devoirs}} submissions, there should be a "grading
directory" set up on your laptop. (You can have more than one, for instance, if you have several sections
that you want to grade separately.)
}
\details{
The grading directory is an ordinary directory that has two requirements:
\enumerate{
\item a \code{course-parameters.yaml} file in the appropriate format
\item a \code{Score-files} subdirectory.
}

\code{create_grading_directory()} will template that minimal structure. Once created, you must then
edit the \code{course-parameters.yaml} file.
a. the \verb{course-name:} will already be filled in (and will also be the name of the directory)
b. the \verb{class-list:} will have dummy names. You need to replace these with the email addresses
that your students will use on the submission site. (You can also add alias emails for any student separated by
spaces in the same line as the student's "official" email. These get sorted out later.)
c. the \verb{submissions-file:} field should be edited to contain the URL to a CSV file on the collection site. How this
is arranged depends on the kind of collection site. See the vignette.

It's recommended that you keep \code{course_name} short and avoid any white space or
characters that are disallowed in file or directory names. Example: "Stats_101_Fall_2025".
}
