% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devoirs-text-entry.R
\name{devoirs_text}
\alias{devoirs_text}
\title{Create a text-entry widget for collection by devoirs package}
\usage{
devoirs_text(
  qID,
  rows = 4,
  cols = 50,
  code = NULL,
  str = "Enter your response to {code} here"
)
}
\arguments{
\item{qID}{unique text string}

\item{rows}{number of lines in empty text area (4)}

\item{cols}{number of characters available in each line (50)}
}
\description{
Create a text-entry widget for collection by devoirs package
}
