% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grading.R
\name{get_new_submissions}
\alias{get_new_submissions}
\alias{get_historic_data}
\alias{update_submissions}
\title{Get the new submissions from the repo}
\usage{
get_new_submissions(home = ".")

get_historic_data(home = ".", since = "2000-1-1 00:00:01 UTC")

update_submissions(home = ".")
}
\arguments{
\item{since}{date for earliest submission to include. Format: "2024-11-18 00:00:01 UTC"}

\item{new_only}{If TRUE, just return the data newly read from the repo.}
}
\description{
Get the new submissions from the repo

Get the historic data

Read the raw submissions data from the repo submissions file and add it
to the Permanent_store.RDS file.
}
